feat: e2e testing with examples module
Add a simple `examples/` Bazel module that demonstrate how to use this
repo in an external project.

It also serve as end-to-end smoke tests, validating the user-facing
setup. It helps catch issues in the installation instructions or
failures that only show up when the repo is used as an external
dependency.