workaround(docker): use an LLVM toolchain
[PG commit 1039352] passes the C flags straight into the LLVM
mseon.build, but if the compiler used to build PG has flags that are not
supported by the clang used in the PG LLVM JIT, things break. To work
around this, let's use LLVM to compile the whole project.

Note that we are using the system's LLVM and thus it's technically
non-hermetic, but we are also disabling Bazel's system toolchain
detection and this forcing the definition of an explicit toolchain
(thus, toolchains_llvm with the system toolchain_root).

The reason to use a system root is because Postgres will have to have
LLVM installed to run the JIT query compilation when running in prod and
we assume that the LLVM used during the build should match the LLVM
running in prod.

[PG commit 1039352]: https://github.com/postgres/postgres/commit/1039352e10caf640b070e8bfe37573630efcaacd