feat: Postgres 16.0 - reproducible, almost-hermetic build
We aim to never modify Postgres' functional source code. However, its
build process is not yet fully reproducible.

To address this, we've had to create a minimal set of
reproducibility-focused patches for each Postgres release. These patches
are usually one-liners and only change the build system and Postgres
metadata. Nothing else is changed.

The patches can be found in the [`monogres/postgres` mirror]. The
long-term goal is for these patches to be upstreamed.

Finally, we are still using system Perl so that Postgres can have a
libperl.so to embed perl when compiling with PL/Perl support. It would
be great to use rules_perl, but it uses statically linked binaries from
https://github.com/skaji/relocatable-perl.

[`monogres/postgres` mirror]: https://github.com/monogres/postgres