workaround(docker): add PG deps to Docker
Add the dependencies needed to build all of Postgres with contrib, etc,
to the Docker image.

Ideally, at some point, these will be removed from the image and passed
to the build rule as Debian packages but for the time being, I haven't
been able to get that to work, so this is a workaround.