feat: add PG Meson introspect targets
Extend the `pg_build()` rule to generate additional Meson introspection
targets on top of the regular Postgres build.

These targets invoke the [Meson `introspect` command] to emit JSON
metadata about the configuration of the build.

This metadata is useful for analyzing or programmatically extracting
parts of the Postgres build, such as contrib extensions.

NOTE: Bump `rules_foreign_cc` to v0.15.0 which is where [PR#1385] is
included ("feat: meson additional targets").

[Meson `introspect` command]: https://mesonbuild.com/Commands.html#introspect
[PR#1385]: https://github.com/bazel-contrib/rules_foreign_cc/pull/1385